Deleting base-package, move AbstractEntities
