+++ src/main/java/com/school/project/model/entities/Attendance.java
@@ -1,6 +1,6 @@
 package com.school.project.model.entities;
 
-import com.school.project.model.base.AbstractBaseEntity;
+import com.school.project.model.AbstractBaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
